Added filter for mars-viewer
Calendar functional for desktop only on Firefox. Functional for both desktop and mobile on Google Chrome
