+++ src/components/marsPhotoViewer/marsViewer.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import "./marsViewer.scss";
+import { ImageFilter } from "./ImageFilter";
 
 export const MarsViewer: React.FunctionComponent = () => {
   return (
@@ -7,6 +8,7 @@ export const MarsViewer: React.FunctionComponent = () => {
       <h1 className="photo-viewer-header">
         Mars Photo Viewer
       </h1>
+      <ImageFilter/>
     {/* 
     Insert the following three components 
     <ImageViewer />
